requirements.txt
Flask==3.0.3
flet==0.22.1
opencv-python==4.10.0.84
paho-mqtt==2.1.0
pillow==10.3.0
mss==10.0.0
pycaw==20240210
pystray==0.19.5
plyer==2.1.0
pywin32==307
keyboard==0.13.5
screeninfo==0.8.1
pywebview==5.3.2